// ClientAdvisor/App/frontend/src/types/User.ts
// User.ts
export interface User {
    ClientId: number;
    ClientName: string;
    ClientEmail: string;
    NextMeeting: string;
    NextMeetingTime: string;
    NextMeetingEndTime: string;
    AssetValue: string;
    LastMeeting: string;
    LastMeetingStartTime: string;
    LastMeetingEndTime: string;
    ClientSummary: string;
    chartUrl: string;

  }
  